+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Application configuration file (.ogma)
+- Logging (ogma.log)
+- New commands
+    - _Add_
+    - _Import_
+    - _Issues_
+    - _Listings_
+- New _info_ flag for base command
+### Changed
+- More unit testing
+
 
 ## [0.0.1] - 2021-10-21
 ### Added